ZC-7277: only do cleanup on erase

+++ SPECS/ea-mod_security2-rules-owasp.spec
@@ -140,8 +140,11 @@ fi
 
 %postun
 
-/usr/local/cpanel/scripts/modsec_vendor remove OWASP3
-rm -rf /var/cpanel/modsec_vendors/meta_OWASP3.cache
+if [ $1 -eq 0 ] ; then
+    echo "Removing OWASP3 config"
+    /usr/local/cpanel/scripts/modsec_vendor remove OWASP3
+    rm -rf /var/cpanel/modsec_vendors/meta_OWASP3.cache
+fi
 
 %files
 %defattr(-, root, root, -)
